move functions that grab tsg errors (bc of cyclical dependencies), change github access errors to warnings
